moved backupchoice to prompt
