+++ packages/persistence/src/backup/index.ts
@@ -1,4 +1,3 @@
-export * from "@persistence/backup/BackupChoice.js";
 export * from "@persistence/backup/BackupManager.js";
 export * from "@persistence/backup/BackupManagerProvider.js";
 export * from "@persistence/backup/IBackupManager.js";
